Update RocketDAOProtocolSettingsInflation.sol
Line 20 & 21 of the RocketDAOProtocolSettingsInflation.sol uses 6,170 blocks as being equal to 1 day. This assumes an average of 14 seconds per block. I did notice that over the last couple of years the average block time was shorter (13.1 sec) . See  https://etherscan.io/chart/blocktime.

With the likelihood of the quick merge coming this year it might be good just to use the eth2 fixed slot time (12 sec) as a good approximation of the current eth1 block time (~13.2 sec).  It will save time adjusting it after the merge. Assuming you went with 12 sec/block the number of block in a day = 7,200. 

Might mean a little few RPL rewards for me as a node operator as the daily inflation would take 26.4 hrs (@7,200) vs 22.5 hrs (@ 6,140) but figured I'd mention this as a suggestion. 

Also, a year in the code is arbitrarily set to 365 days, not 365.25